requirements.txt
requests>=2.28.0
PyGithub>=1.59.0
GitPython>=3.1.30
Flask>=2.0.0
gunicorn>=20.1.0
cryptography>=3.4.0
